requirements.txt
streamlit==0.52.2
tweepy==3.8.0
vaderSentiment==3.2.1
wordcloud==1.6.0
pandas==0.25.3
spacy==2.2.3
nltk==3.4.5
plotly==4.4.1
matplotlib==3.1.2
gensim==3.8.1
pyLDAvis==2.1.2
Pillow==6.2.1
textblob==0.15.3
sklearn==0.0
wordcloud==1.6.0




